+++ sdk/keyvault/azure-keyvault-keys/tests/test_challenge_auth.py
@@ -7,6 +7,7 @@ Tests for the HTTP challenge authentication implementation. These tests aren't p
 the challenge cache is global to the process.
 """
 import functools
+import os
 import time
 from uuid import uuid4
 
@@ -20,11 +21,44 @@ from azure.core.exceptions import ServiceRequestError
 from azure.core.pipeline import Pipeline
 from azure.core.pipeline.policies import SansIOHTTPPolicy
 from azure.core.pipeline.transport import HttpRequest
+from azure.identity import ClientSecretCredential
+from azure.keyvault.keys import KeyClient
 from azure.keyvault.keys._shared import ChallengeAuthPolicy, HttpChallenge, HttpChallengeCache
 
 import pytest
 
 from _shared.helpers import mock_response, Request, validating_transport
+from _shared.test_case import KeyVaultTestCase
+from _test_case import client_setup, get_decorator, KeysTestCase
+
+
+all_api_versions = get_decorator()
+
+
+class ChallengeAuthTests(KeysTestCase, KeyVaultTestCase):
+    def __init__(self, *args, **kwargs):
+        super(ChallengeAuthTests, self).__init__(*args, match_body=False, **kwargs)
+
+    @all_api_versions()
+    @client_setup
+    def test_multitenant_authentication(self, client, is_hsm, **kwargs):
+        client_id = os.environ.get("KEYVAULT_CLIENT_ID")
+        client_secret = os.environ.get("KEYVAULT_CLIENT_SECRET")
+        if self.is_live and not client_id and client_secret:
+            pytest.skip("Values for KEYVAULT_CLIENT_ID and KEYVAULT_CLIENT_SECRET are required")
+
+        # we set up a client for this method to align with the async test, but we actually want to create a new client
+        # this new client should use a credential with an initially fake tenant ID and still succeed with a real request
+        api_version = client.api_version
+        credential = ClientSecretCredential(tenant_id=str(uuid4()), client_id=client_id, client_secret=client_secret)
+        vault_url = self.managed_hsm_url if is_hsm else self.vault_url
+        client = KeyClient(vault_url=vault_url, credential=credential, api_version=api_version)
+
+        if self.is_live:
+            time.sleep(2)  # to avoid throttling by the service
+        key_name = self.get_resource_name("multitenant-key")
+        key = client.create_rsa_key(key_name)
+        assert key.id
 
 
 def empty_challenge_cache(fn):